refactored recyclerview to main activity
