+++ app/src/main/res/layout/activity_upgrade_menu.xml
@@ -6,18 +6,4 @@
     android:layout_height="match_parent"
     tools:context=".activity.UpgradeMenuActivity">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/upgrade_menu_list_recycler_view"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="100dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginBottom="100dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0" />
 </androidx.constraintlayout.widget.ConstraintLayout>
